requirements.txt
amqp==1.4.9
anyjson==0.3.3
awsebcli==3.7.5
billiard==3.3.0.23
blessed==1.9.5
boto==2.39.0
botocore==1.4.10
celery==3.1.23
cement==2.8.2
colorama==0.3.7
Django==1.9.5
django-celery==3.1.17
django-storages==1.4.1
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
httplib2==0.9.2
jmespath==0.9.0
kombu==3.0.35
pathspec==0.3.4
Pillow==3.2.0
psycopg2==2.6.1
PySocks==1.5.6
python-dateutil==2.5.2
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
twilio==5.4.0
uuid==1.30
wcwidth==0.1.6
websocket-client==0.35.0
